+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 2.9.9
 - Fix SQL error when more than one Rollover default enrollment code in Rollover.php
 - Add warning & help for Rollover default enrollment code in EnrollmentCodes.php
 - Fix Delete Prompt displayed when working User/Student cleared program wide
+- Add Warning if not in current Quarter in Assignments.php
 
 Changes in 2.9.8
 ----------------